requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
pymongo==4.11.0
dnspython==2.7.0
chromadb==0.4.22
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.1
pysqlite3==0.5.2
matplotlib==3.8.3
